feat(resolver): rework yarn manifest file look up
The current manifest file (`.pnp.cjs`) lookup technique does not work when
`enableGlobalCache` is set to `true`, which is the default behaviour.

Instead, we try and find the closest pnp maniest from cwd.

This will break the multi-project case mentioned in https://yarnpkg.com/advanced/pnp-spec#find_pnp_manifest,
but I think this is an acceptable trade off.