feat(resolver): rework yarn manifest file look up (#586)
The current manifest file (`.pnp.cjs`) lookup technique does not work
when `enableGlobalCache` is set to `true`, which is the default behavior
for yarn.

Instead, we try and find the closest pnp maniest from cwd.

This will break the multi-project case mentioned in
https://yarnpkg.com/advanced/pnp-spec#find_pnp_manifest, but I think
this is an acceptable trade off.

For napi, I set `process.env.OXC_RESOLVER_YARN_PNP` in `napi/index.js`
so the Rust side can enable yarn pnp accordingly.