fix: do not resolve to `node_modules/pacakge/index` (#849)
fixes #845

The option `main_files` is kind of confusing, it's supposed to be the file stem without the extension (e.g. index), but somehow it ended up being a lookup for `index` then `index.js` instead of just `index.js`.

This reduces an extra "is file" system call for `node_modules/package/index`.